// packages/@mocktailgpt/ts/src/generated/models/chatCompletion.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * van-scrapper API
 * API for retrieving, analyzing and structuring van info for van conversion business, resale, and homologation in Belgium. All endpoints POST, OpenAI completions format only. All properties in English, string values in source language (fr/en). Enums factorized.

 * OpenAPI spec version: 2.0
 */
import type { ChatCompletionObject } from './chatCompletionObject';
import type { ChatCompletionChoice } from './chatCompletionChoice';

export interface ChatCompletion {
  id: string;
  object: ChatCompletionObject;
  created: number;
  model: string;
  choices: ChatCompletionChoice[];
}